+++ synnax/pkg/service/hardware/device/device.go
@@ -12,7 +12,7 @@
 package device
 
 import (
-	"github.com/synnaxlabs/synnax/pkg/hardware/rack"
+	"github.com/synnaxlabs/synnax/pkg/service/hardware/rack"
 	"github.com/synnaxlabs/x/gorp"
 	"github.com/synnaxlabs/x/validate"
 )
