+++ synnax/pkg/service/hardware/task/task.go
@@ -16,7 +16,7 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/synnaxlabs/synnax/pkg/hardware/rack"
+	"github.com/synnaxlabs/synnax/pkg/service/hardware/rack"
 	"github.com/synnaxlabs/x/errors"
 	"github.com/synnaxlabs/x/gorp"
 	"github.com/synnaxlabs/x/types"
