+++ synnax/pkg/service/workspace/lineplot/retrieve_test.go
@@ -12,7 +12,7 @@ package lineplot_test
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/synnaxlabs/synnax/pkg/workspace/schematic"
+	"github.com/synnaxlabs/synnax/pkg/service/workspace/schematic"
 )
 
 var _ = Describe("Retrieve", func() {
