+++ synnax/pkg/service/workspace/schematic/retrieve_test.go
@@ -12,7 +12,7 @@ package schematic_test
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/synnaxlabs/synnax/pkg/workspace/schematic"
+	"github.com/synnaxlabs/synnax/pkg/service/workspace/schematic"
 )
 
 var _ = Describe("Retrieve", func() {
